% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_canopy}
\alias{boston_canopy}
\title{Boston tree canopy change (2014-2019) and heat index}
\format{
An sf object TK
}
\usage{
boston_canopy
}
\description{
A dataset containing TK
}
\keyword{datasets}
