% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_indices}
\alias{buffer_indices}
\title{Apply an inclusion radius and exclusion buffer to indices}
\usage{
buffer_indices(data, indices, radius, buffer, call = rlang::caller_env())
}
\arguments{
\item{data}{An object of class \code{sf} or \code{sfc}.}

\item{indices}{List of indices in each fold generated by \code{split_unnamed()}.}

\item{radius}{Numeric: points within this distance of the initially-selected
test points will be assigned to the assessment set.}

\item{buffer}{Numeric: points within this distance of the test set
(after \code{radius} is applied)
will be assigned to neither the analysis or assessment set.}
}
\description{
Apply an inclusion radius and exclusion buffer to indices
}
\keyword{internal}
