% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_v}
\alias{check_v}
\title{Check that "v" is sensible}
\usage{
check_v(v, max_v, objects, allow_max_v = TRUE, call = rlang::caller_env())
}
\arguments{
\item{v}{The number of partitions for the resampling. Set to \code{NULL} or \code{Inf}
for the maximum sensible value (for leave-one-X-out cross-validation).}
}
\description{
Check that "v" is sensible
}
\keyword{internal}
