% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_location_out_cv.R
\name{leave_location_out_cv}
\alias{leave_location_out_cv}
\title{Leave-Location-Out Cross Validation}
\usage{
leave_location_out_cv(
  data,
  location,
  v = length(unique(location)),
  pool = 0,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{location}{A variable in data (single character or name) used to
indicate "location" groupings. Data are immediately converted to factors,
so it is possible to use integer indicators.}

\item{v}{The number of partitions of the data set.}

\item{pool}{A proportion of data used to determine if a particular location
is too small and should be pooled into another group. For instance, if
\code{pool = 0.1}, then only locations constituting 10\% or more of the data will
be preserved, with the remainder being folded into a single "pooled" location.}

\item{...}{Not currently used.}
}
\description{
In situations where you have a clear idea of what data are most likely
to be related to one another, it often makes sense to use those groupings
to create validation folds rather than trying to cluster or block data into
sensible groups.
For instance, if your samples are collected from several islands and you're
primarily interested in how your model generalizes to unmeasured islands,
then the easiest way to assess your model is to treat each island as a "fold"
for the purposes of v-fold CV.
This function breaks \code{data} into \code{v} folds, grouping data which share a
\code{location} into the same folds. When \code{v} is equal to the number of unique
\code{location}s, this function produces "leave-one-location-out" resamples.
}
\references{
H. Meyer, C. Reudenbach, T. Hengl, M. Katurji, and T. Nauss, 2018.
"Improving performance of spatio-temporal machine learning models using
forward feature selection and target-oriented validation,"
Environmental Modelling & Software 101,
pp. 1-9, doi: 10.1016/j.envsoft.2017.12.001.
}
