% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{pretty.spatial_clustering_cv}
\alias{pretty.spatial_clustering_cv}
\title{Short Descriptions of spatial rsets}
\usage{
\method{pretty}{spatial_clustering_cv}(x, ...)
}
\arguments{
\item{x}{An \code{rset} object}

\item{...}{Not currently used.}
}
\value{
A character vector.
}
\description{
Produce a character vector describing the spatial resampling method.
}
\keyword{internal}
