% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{analysis}
\alias{assessment}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rsample}{\code{\link[rsample:as.data.frame.rsplit]{analysis}}, \code{\link[rsample:as.data.frame.rsplit]{assessment}}}
}}

