% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{spatialsample-package}
\alias{spatialsample}
\alias{spatialsample-package}
\title{spatialsample: Spatial Resampling Infrastructure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions and classes for spatial resampling to use with the 'rsample' package, such as spatial cross-validation (Brenning, 2012) \doi{10.1109/IGARSS.2012.6352393}. The scope of 'rsample' and 'spatialsample' is to provide the basic building blocks for creating and analyzing resamples of a spatial data set, but neither package includes functions for modeling or computing statistics. The resampled spatial data sets created by 'spatialsample' do not contain much overhead in memory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/spatialsample}
  \item \url{https://spatialsample.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/spatialsample/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@rstudio.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

Authors:
\itemize{
  \item Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
