% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.spatial_rset}
\alias{autoplot.spatial_rset}
\title{Create a ggplot for spatial resamples.}
\usage{
\method{autoplot}{spatial_rset}(object, ...)
}
\arguments{
\item{object}{A \code{spatial_rset} object. Note that only resamples made from
\code{sf} objects are \code{spatial_rset} objects; this function will not work for
resamples made with non-spatial tibbles or data.frames.}

\item{...}{Options passed to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.}
}
\value{
A ggplot object with each fold assigned a color, made using
\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.
}
\description{
This method provides a good visualization method for spatial resampling.
}
\examples{
\dontshow{if (rlang::is_installed("ggplot2") && rlang::is_installed("modeldata")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(ames, package = "modeldata")
ames_sf <- sf::st_as_sf(ames,
                        coords = c("Longitude", "Latitude"), crs = 4326)

ames_block <- spatial_block_cv(ames_sf)
autoplot(ames_block)
\dontshow{\}) # examplesIf}
}
