% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_indices}
\alias{buffer_indices}
\title{Apply an inclusion radius and exclusion buffer to indices}
\usage{
buffer_indices(data, indices, radius, buffer, ...)
}
\arguments{
\item{data}{An object of class \code{sf} or \code{sfc}.}

\item{indices}{List of indices in each fold generated by \code{split_unnamed}.}

\item{radius}{Numeric: points within this distance of the initially-selected
test points will be assigned to the assessment set.}

\item{buffer}{Numeric: points within this distance of the test set area
(that is, the area within \code{radius} distance of the initial test points)
will be assigned to neither the analysis or assessment set.}

\item{...}{Arguments passed to \code{\link[sf:geos_unary]{sf::st_buffer()}}. Note that both the \code{radius}
and \code{buffer} zones are created using \code{\link[sf:geos_unary]{sf::st_buffer()}}, so any arguments
passed through \code{...} will be applied to both.}
}
\description{
Apply an inclusion radius and exclusion buffer to indices
}
