% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leave_location_out_cv.R
\name{make_location}
\alias{make_location}
\title{Create or Modify Location Variables}
\usage{
make_location(x, pool = 0.1)
}
\arguments{
\item{x}{An input vector.}

\item{pool}{A proportion of data used to determine if a particular group is
too small and should be pooled into another group.}
}
\value{
A factor vector.
}
\description{
Create or Modify Location Variables
}
\details{
For categorical inputs, the function will find levels of \code{x} than
occur in the data with percentage less than \code{pool}. The values from
these groups will be pooled into a single category (if they make up
a greater than \code{pool} percentage of the data) or combined with the smallest
extant category (if not).
}
\keyword{internal}
