% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_block_cv.R, R/spatial_clustering_cv.R
\name{spatial_block_cv}
\alias{spatial_block_cv}
\alias{spatial_clustering_cv}
\title{Spatial Cross-Validation}
\usage{
spatial_block_cv(
  data,
  coords,
  v = 10,
  nrow = 5,
  ncol = 5,
  crs = 4326,
  outcome = "",
  allocation = c("random", "systematic", "checkboard"),
  save_blocks = FALSE,
  verbose = FALSE,
  progress = FALSE,
  show_blocks = FALSE,
  ...
)

spatial_clustering_cv(
  data,
  coords,
  v = 10,
  method = c("kmeans", "kmedoids"),
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{coords}{A vector of variable names, typically spatial coordinates, to
partition the data into disjointed sets via k-means clustering.}

\item{v}{The number of partitions of the data set.}

\item{...}{Extra arguments passed on to clustering functions,
\code{\link[stats:kmeans]{stats::kmeans()}} for kmeans, and.}

\item{method}{A character string for clustering algorithm (i.e., "kmeans" or
"kmedoids")}
}
\value{
A tibble with classes \code{spatial_cv}, \code{rset}, \code{tbl_df}, \code{tbl}, and
\code{data.frame}. The results include a column for the data split objects and
an identification variable \code{id}.
}
\description{
Spatial cross-validation splits the data into V groups of disjointed sets.
These sets are separated using spatial coordinates such that observations
within the set are geographically closer to each other. This is helpful to
reduce spatial autocorrelation compared to random cross validation.
}
\details{
The variables in the \code{coords} argument are used for k-means
clustering of the data into disjointed sets, as outlined in Brenning (2012).
These clusters are used as the folds for cross-validation. Depending on how
the data are distributed spatially, there may not be an equal number of
points in each fold.

\link{spatial_clustering_cv} uses clustering algorithms to create spatial
clusters. A resample of the analysis data consists of V-1 of the
folds/clusters while the assessment set contains the final fold/cluster. In
basic spatial cross-validation (i.e. no repeats), the number of resamples is
equal to V.

\link{spatial_block_cv} makes partitions with spatial blocks or buffers and
allocate them into cross validation folds. Blocks are units of geographical
area (e.g. rectangles, spatial polygons and buffers of specific distance)
Within these units, all species data are treated together—for instance,
allocated to the same fold of cv. Several blocks could be allocated to one cv
fold.
}
\examples{
data(ames, package = "modeldata")
spatial_clustering_cv(ames, coords = c(Latitude, Longitude), v = 5)

}
\references{
Valavi, R, Elith, J, Lahoz-Monfort, JJ, Guillera-Arroita, G. blockCV:
An r package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models.
Methods Ecol Evol. 2019; 10: 225– 232.
https://doi.org/10.1111/2041-210X.13107

A. Brenning, "Spatial cross-validation and bootstrap for the assessment of
prediction rules in remote sensing: The R package sperrorest," 2012 IEEE
International Geoscience and Remote Sensing Symposium, Munich, 2012, pp.
5372-5375, doi: 10.1109/IGARSS.2012.6352393.
}
