% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_clustering_cv.R
\name{spatial_clustering_cv}
\alias{spatial_clustering_cv}
\alias{spatial_clustering_cv.default}
\alias{spatial_clustering_cv.sf}
\title{Spatial or Cluster Cross-Validation}
\usage{
spatial_clustering_cv(data, coords, v = 10, fun = c("kmeans", "hclust"), ...)

\method{spatial_clustering_cv}{default}(data, coords, v = 10, fun = c("kmeans", "hclust"), ...)

\method{spatial_clustering_cv}{sf}(data, coords, v = 10, fun = c("kmeans", "hclust"), ...)
}
\arguments{
\item{data}{A data frame, or an \code{sf} object (often from \code{\link[sf:st_read]{sf::read_sf()}}
or \code{\link[sf:st_as_sf]{sf::st_as_sf()}}), to split into folds.}

\item{coords}{A vector of variable names, typically spatial coordinates,
to partition the data into disjointed sets via k-means clustering.
This argument is ignored (with a warning) if \code{data} is an \code{sf} object.}

\item{v}{The number of partitions of the data set.}

\item{fun}{Which function to use to cluster. Must be one of either
"kmeans" (to use \code{\link[stats:kmeans]{stats::kmeans()}}) or "hclust" (to use \code{\link[stats:hclust]{stats::hclust()}}).}

\item{...}{Extra arguments passed on to \code{\link[stats:kmeans]{stats::kmeans()}} or
\code{\link[stats:hclust]{stats::hclust()}}.}
}
\value{
A tibble with classes \code{spatial_cv}, \code{rset}, \code{tbl_df}, \code{tbl}, and
\code{data.frame}. The results include a column for the data split objects and
an identification variable \code{id}.
}
\description{
Spatial or cluster cross-validation splits the data into V groups of
disjointed sets using k-means clustering of some variables, typically
spatial coordinates. A resample of the analysis data consists of V-1 of the
folds/clusters while the assessment set contains the final fold/cluster. In
basic spatial cross-validation (i.e. no repeats), the number of resamples
is equal to V.
}
\details{
The variables in the \code{coords} argument are used for k-means clustering of
the data into disjointed sets, as outlined in Brenning (2012), or for
hierarchical clustering of the data (by default using the complete linkage
method, though this can be changed through arguments). These
clusters are used as the folds for cross-validation. Depending on how the
data are distributed spatially, there may not be an equal number of points
in each fold.
}
\examples{
data(Smithsonian, package = "modeldata")
spatial_clustering_cv(Smithsonian, coords = c(latitude, longitude), v = 5)

# When providing sf objects, coords are inferred automatically
smithsonian_sf <- sf::st_as_sf(Smithsonian,
                        coords = c("longitude", "latitude"),
                        crs = 4326) # WGS84
spatial_clustering_cv(smithsonian_sf, v = 5)

}
\references{
A. Brenning, "Spatial cross-validation and bootstrap for the assessment of
prediction rules in remote sensing: The R package sperrorest," 2012 IEEE
International Geoscience and Remote Sensing Symposium, Munich, 2012,
pp. 5372-5375, doi: 10.1109/IGARSS.2012.6352393.
}
